+++ app/db/errors.py
@@ -1,0 +1,5 @@
+"""Database layer exceptions."""
+
+
+class MigrationChecksumError(RuntimeError):
+    """Raised when an already applied migration script was modified."""